requirements.txt
absl-py==0.15.0
appdirs==1.4.4
astunparse==1.6.3
audioread==2.1.9
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coloredlogs==15.0.1
cycler==0.11.0
decorator==5.1.0
flatbuffers==2.0
future==0.18.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
h5py==3.5.0
humanfriendly==10.0
idna==3.3
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
librosa==0.8.1
llvmlite==0.37.0
Markdown==3.3.4
matplotlib==3.4.3
mir-eval==0.6
numba==0.54.1
numpy==1.20.3
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.2
pandas==1.3.4
Pillow==8.4.0
pooch==1.5.2
progressbar2==3.55.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyFFTW==0.12.0
pyparsing==3.0.5
pyrubberband==0.3.0
PySoundFile==0.9.0.post1
python-dateutil==2.8.2
python-utils==2.5.6
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
scikit-learn==1.0.1
scipy==1.7.2
six==1.16.0
SoundFile==0.10.3.post1
sox==1.4.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.21.0
termcolor==1.1.0
threadpoolctl==3.0.0
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
